Do not require settings declaration if server support setting-as-string
New settings are added frequently, more then clickhouse-driver has new
release.

The downside is that bool type detection now depends from the server,
and it does not recognize 'yes'/'2', but recognize 'true'/'1', but I
guess it is ok.
(plus I'm not sure that it was a good idea to ignore case for char
settings, if the string has more then 1 char)